refactor: enum 클래스를 constant 패키지로 이동
