+++ src/test/java/domain/BlackJackResultTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.constant.GamerResult;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
