+++ src/test/java/domain/CardNumberTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.constant.CardNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
