+++ src/test/java/domain/CardTypeTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.constant.CardType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
