+++ src/test/java/domain/HandTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.constant.CardNumber;
+import domain.constant.CardType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
